Fix bug that caused the server to send null instead of the correct data.
The problem was with the asynchronus nature of Node.js. The server sent the response after clearing the cached list, but before filling it up again.
Unfortunately, the solution wasn't as straightforward as it may seem, because the mongoose's .exec after a query can't be awaited, so I had to turn
two functions into promises, which fixed the problem, because now I was able to simply use .then after calling getHotSmokin and updateHotSmokin.

TODO: debug log