Add config.js file
If the content of the config.js file is in the README.md file anyways, there's no point in excluding it from the repository's files.

Also, the database url has been replaced with an url to the dev database, so the previously inputted data cannot be accessed anymore.
This will make the transition to a new database easier once the website goes live.