Make user-specific routes login-protected and add redirection to router
The watchlist and the account settings page now redirect to the login page if the user is not logged in.

The router now has an addRedirect middleware, that adds a redirect function to the response object, which sends a 302 redirect when called. This is automatically enabled by setup if not specified otherwise.

The server now automatically responds with the preferred page (watchlist/main) when a logged in user is at '/'.