Modify folder structure
The root folder is now the npm base folder (this where the package.json and package-lock.json files are).

This means that instead of starting the server from the src/ folder it now needs to be started from the root folder.

Also, the config.js file has been removed and added to .gitignore, and a config.default.js file has been added with placeholder values.