Change auth-token cookie SameSite directive to Lax
auth-token cookies used to have their SameSite directive set to Strict.
This blocked browsers from sending this cookie to the server when the user opened the site from a link on an external page (e.g. Facebook).
Changing this to Lax means logged in users will get the signed in version of the page when it's opened from an external link.