Add watchlist latency setting and changing
The watchlist latency represents the number of minutes the page will use to pick out relevant trips.

This value can be set during signup (POST /api/users) with the watchlistLatency body field.
Later, this value can be modified the same way as above, but with the account settings modifying route (PUT /api/users/username).

This latency can be accessed in two ways:
- directly, with the user's account settings route (GET /api/users/username)
- the watchlist also contains this information for convenience (GET /api/users/username/watchlist)