Optimize main page watchlist loading
Main page has been changes, so that the user's watchlist is only
loaded when the page loads, and instead of using the real watchlist to
determine whether a segment is in the wathclist or not a local copy is
used, that's being updated locally everytime a change is made to the
watchlist. This makes choosing segments when logged in faster.